+++ Changes.md
@@ -2,7 +2,7 @@
 
 ## Version 1.9
 
-- Supports controls bound to the POV2 hat
+- Supports controls bound to the POV2 hat and throttle
 - Sets all lights to the inactive state even if the are not mapped to recognised
   controls; this is noticable for example when hardpoints are deployed using the
   default configuration